requirements.txt
aiohttp==3.5.4
GitPython==2.1.11
python-etcd==0.4.5
ruamel.yaml==0.15.100
docker==4.0.2
docker-pycreds==0.4.0
flask-swagger==0.2.12
